sysfs: Add /sys/kernel/realtime entry
Add a /sys/kernel entry to indicate that the kernel is a
realtime kernel.

Clark says that he needs this for udev rules, udev needs to evaluate
if its a PREEMPT_RT kernel a few thousand times and parsing uname
output is too slow or so.

Are there better solutions? Should it exist and return 0 on !-rt?

Backported from the PREEMPT_RT v6.18.35-rt5 patch series. Only the 8250
nbcon console and sysfs patches from that series are carried on this
branch; the remaining rt5 patches are ARM, POWERPC and i915 specific and
do not apply to RISC-V.

Signed-off-by: Clark Williams <williams@redhat.com>
Signed-off-by: Peter Zijlstra <a.p.zijlstra@chello.nl>
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Mike Maizelis <mike.maizelis@microchip.com>