trace: prevent osnoise hotplog worker running in UNBOUND workqueue
smp_processor_id() should be called with migration disabled. This mean
we may safely call smp_processor_id() in percpu thread. However, this is
not the case if the work is (re-)queued into unbound workqueue, during
cpu-hotplog. So, detect and return early if this work happens to run on
an unbound wq.

Signed-off-by: Andy Chiu <andy.chiu@sifive.com>