riscv: mm: execute local TLB flush after populating vmemmap
The vmemmap_populate() creates VA to PA mapping for the VMEMMAP area, where
all "strcut page" are located once CONFIG_SPARSEMEM_VMEMMAP is defined.
These "struct page" are later initialized in the zone_sizes_init()
function. However, during this process, no sfence.vma instruction is
executed for this VMEMMAP area. This omission may cause the core to fail to
perform page table work because some data related to the address
translation is invisible to the core. To solve this issue, here let
zone_sizes_init() call local_flush_tlb_kernel_range() to execute a
sfence.vma instruction for the VMEMMAP area, ensuring that all data related
to the address translation is visible to the core.

Signed-off-by: Vincent Chen <vincent.chen@sifive.com>