sched_clock: Support the sched_clock devices with rates exceeding 4G.
The sched_clock_register function cannot support devices with clock
frequencies exceeding 4GHz. To address this issue, we refer to the
implementation of clocksource_register_khz and introduce a new interface
called clocksource_register_khz. This enables drivers to register a
sched_clock device with a clock frequency exceeding 4GHz, resolving the
problem.

Signed-off-by: Vincentc Chen <vincent.chen@sifive.com>