ARM: dts: microchip: sama7d65eb: Introduce local macro to select LCD I/F
Introduce a local macro - LCD_DSI_NOT_LVDS - for the preprocessor to
compile and enable either one of the two LCD interfaces featured on the
board.

Have this macro defined for MIPI DSI, or have it undefined for LVDS.

Conflicts of multiple endpoints pointing to a same endpoint shall be
prevented even if the node status is disabled.

Signed-off-by: Aubin Constans <aubin.constans@microchip.com>
Acked-by: Ryan Wanner <ryan.wanner@microchip.com>